exercise/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
distlib==0.3.1
dj-database-url==0.5.0
Django==2.2.13
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-multiselectfield==0.1.12
djangorestframework==3.11.0
filelock==3.0.12
importlib-metadata==1.7.0
importlib-resources==3.0.0
Pillow==7.2.0
psycopg2==2.8.6
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
virtualenv==20.0.31
whitenoise==5.2.0
zipp==3.1.0
